fix lab 7

+++ src/main/java/ru/ahmetdavresh/springthymeleafwebapp/dao/StudentRepository.java
@@ -3,5 +3,5 @@ package ru.ahmetdavresh.springthymeleafwebapp.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.ahmetdavresh.springthymeleafwebapp.entity.Student;
 
-public interface StudentRepository extends JpaRepository<Student, Long> {
+public interface StudentRepository extends JpaRepository<Student, Integer> {
 }
